requirements.txt
aiodns==3.2.0
aiohttp==3.9.5
aiosignal==1.3.1
anyio==4.4.0
attrs==23.2.0
backoff==2.2.1
Backtesting==0.3.3
bitarray==2.9.2
bokeh==3.4.2
ccxt==4.3.55
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
ckzg==1.0.2
contourpy==1.2.1
cryptography==42.0.8
cytoolz==0.12.3
eth-account==0.11.2
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.5.1
eth-rlp==1.0.1
eth-typing==4.3.1
eth-utils==4.1.1
eth_abi==5.1.0
frozenlist==1.4.1
gql==3.5.0
graphql-core==3.2.3
hexbytes==0.3.1
idna==3.7
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kwenta==1.2.0
lru-dict==1.2.0
MarkupSafe==2.1.5
multidict==6.0.5
numpy==2.0.0
packaging==24.1
pandas==2.2.2
pandas_ta==0.3.14b0
parsimonious==0.10.0
pillow==10.3.0
protobuf==5.27.2
pycares==4.4.0
pycparser==2.22
pycryptodome==3.20.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pyunormalize==15.1.0
PyYAML==6.0.1
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
rlp==4.0.1
rpds-py==0.18.1
schedule==1.2.2
six==1.16.0
sniffio==1.3.1
toolz==0.12.1
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
web3==6.20.0
websockets==12.0
xyzservices==2024.6.0
yarl==1.9.4
